requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_tflow_select=2.3.0=mkl
absl-py=0.9.0=py37_0
appnope=0.1.0=py37_0
astor=0.8.0=py37_0
attrs=19.3.0=py_0
backcall=0.1.0=py37_0
blas=1.0=mkl
bleach=3.1.0=py37_0
boto3=1.12.26=pypi_0
botocore=1.15.26=pypi_0
c-ares=1.15.0=h1de35cc_1001
ca-certificates=2020.1.1=0
certifi=2019.11.28=py37_0
chardet=3.0.4=pypi_0
click=7.1.1=pypi_0
cycler=0.10.0=py37_0
dbus=1.13.12=h90a0687_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
docutils=0.15.2=pypi_0
entrypoints=0.3=py37_0
expat=2.2.6=h0a44026_0
filelock=3.0.12=pypi_0
freetype=2.9.1=hb4e5f40_0
gast=0.2.2=py37_0
gettext=0.19.8.1=h15daf44_3
glib=2.63.1=hd977a24_0
google-pasta=0.1.8=py_0
grpcio=1.16.1=py37h044775b_1
h5py=2.10.0=py37h3134771_0
hdf5=1.10.4=hfa1e0ec_0
icu=58.2=h4b95b61_1
idna=2.9=pypi_0
importlib_metadata=1.5.0=py37_0
intel-openmp=2019.4=233
ipykernel=5.1.4=py37h39e3cac_0
ipython=7.13.0=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=py_0
jedi=0.16.0=py37_0
jinja2=2.11.1=py_0
jmespath=0.9.5=pypi_0
joblib=0.14.1=py_0
jpeg=9b=he5867d9_2
jsonschema=3.2.0=py37_0
jupyter=1.0.0=py37_7
jupyter_client=6.0.0=py_0
jupyter_console=6.1.0=py_0
jupyter_core=4.6.1=py37_0
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=py37h0a44026_0
libcxx=4.0.1=hcfea43d_1
libcxxabi=4.0.1=hcfea43d_1
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h475c297_4
libgfortran=3.0.1=h93005f0_2
libiconv=1.15=hdd342a3_7
libpng=1.6.37=ha441bb4_0
libprotobuf=3.11.4=hd9629dc_0
libsodium=1.0.16=h3efe00b_0
libtiff=4.1.0=hcb84e12_0
llvm-openmp=4.0.1=hcfea43d_1
markdown=3.1.1=py37_0
markupsafe=1.1.1=py37h1de35cc_0
matplotlib=3.1.3=py37_0
matplotlib-base=3.1.3=py37h9aa3819_0
mistune=0.8.4=py37h1de35cc_0
mkl=2019.4=233
mkl-service=2.3.0=py37hfbe908c_0
mkl_fft=1.0.15=py37h5e564d8_0
mkl_random=1.1.0=py37ha771720_0
nbconvert=5.6.1=py37_0
nbformat=5.0.4=py_0
ncurses=6.2=h0a44026_0
ninja=1.9.0=py37h04f5b5a_0
notebook=6.0.3=py37_0
numpy=1.18.1=py37h7241aed_0
numpy-base=1.18.1=py37h6575580_1
olefile=0.46=py37_0
openssl=1.1.1e=h1de35cc_0
opt_einsum=3.1.0=py_0
pandas=1.0.2=py37h6c726b0_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py37_1
parso=0.6.2=py_0
pcre=8.43=h0a44026_0
pexpect=4.8.0=py37_0
pickleshare=0.7.5=py37_0
pillow=7.0.0=py37h4655f20_0
pip=20.0.2=py37_1
prometheus_client=0.7.1=py_0
prompt_toolkit=3.0.3=py_0
protobuf=3.11.4=py37h0a44026_0
ptyprocess=0.6.0=py37_0
pygments=2.6.1=py_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py37h655552a_2
pyrsistent=0.15.7=py37h1de35cc_0
python=3.7.6=h359304d_2
python-dateutil=2.8.1=py_0
pytorch=1.4.0=py3.7_0
pytz=2019.3=py_0
pyzmq=18.1.1=py37h0a44026_0
qt=5.9.7=h468cd18_1
qtconsole=4.7.1=py_0
qtpy=1.9.0=py_0
readline=7.0=h1de35cc_5
regex=2020.2.20=pypi_0
requests=2.23.0=pypi_0
s3transfer=0.3.3=pypi_0
sacremoses=0.0.38=pypi_0
scikit-learn=0.22.1=py37h27c97d8_0
scipy=1.4.1=py37h9fa6033_0
send2trash=1.5.0=py37_0
sentencepiece=0.1.85=pypi_0
setuptools=46.0.0=py37_0
sip=4.19.8=py37h0a44026_0
six=1.14.0=py37_0
sqlite=3.31.1=ha441bb4_0
tensorboard=2.0.0=pyhb38c66f_1
tensorflow=2.0.0=mkl_py37hda344b4_0
tensorflow-base=2.0.0=mkl_py37h66b1bf0_0
tensorflow-estimator=2.0.0=pyh2649769_0
termcolor=1.1.0=py37_1
terminado=0.8.3=py37_0
testpath=0.4.4=py_0
tk=8.6.8=ha441bb4_0
tokenizers=0.5.2=pypi_0
torchvision=0.5.0=py37_cpu
tornado=6.0.4=py37h1de35cc_1
tqdm=4.43.0=pypi_0
traitlets=4.3.3=py37_0
transformers=2.5.1=pypi_0
urllib3=1.25.8=pypi_0
wcwidth=0.1.8=py_0
webencodings=0.5.1=py37_1
werkzeug=1.0.0=py_0
wheel=0.34.2=py37_0
widgetsnbextension=3.5.1=py37_0
wrapt=1.12.1=py37h1de35cc_1
xlrd=1.2.0=py37_0
xz=5.2.4=h1de35cc_4
zeromq=4.3.1=h0a44026_3
zipp=2.2.0=py_0
zlib=1.2.11=h1de35cc_3
zstd=1.3.7=h5bba6e5_0
